+++ opentelemetry-dynatrace/src/metric.rs
@@ -13,17 +13,17 @@ use http::{
     Method, Uri, Version,
 };
 use opentelemetry::metrics::Result;
-use opentelemetry::runtime::Runtime;
-use opentelemetry::sdk::export::metrics::aggregation::{
-    AggregationKind, Temporality, TemporalitySelector,
-};
-use opentelemetry::sdk::export::metrics::{AggregatorSelector, InstrumentationLibraryReader};
-use opentelemetry::sdk::metrics::controllers::BasicController;
-use opentelemetry::sdk::metrics::sdk_api::Descriptor;
-use opentelemetry::sdk::metrics::{controllers, processors};
-use opentelemetry::sdk::{export::metrics, Resource};
 use opentelemetry::{global, Context};
 use opentelemetry_http::HttpClient;
+use opentelemetry_sdk::export::metrics::aggregation::{
+    AggregationKind, Temporality, TemporalitySelector,
+};
+use opentelemetry_sdk::export::metrics::{AggregatorSelector, InstrumentationLibraryReader};
+use opentelemetry_sdk::metrics::controllers::BasicController;
+use opentelemetry_sdk::metrics::sdk_api::Descriptor;
+use opentelemetry_sdk::metrics::{controllers, processors};
+use opentelemetry_sdk::runtime::Runtime;
+use opentelemetry_sdk::{export::metrics, Resource};
 use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::fmt::{Debug, Formatter, Write};